Write attribute data to NetCDF-CF

Creates a NetCDF file with an instance dimension, and any attributes from a data frame. Use to create the start of a NetCDF-DSG file. One character length dimension is created long enough to contain the longest provided character string. This function does not implement any CF convention attributes or standard names. Any columns of class date will be converted to character.

write_attribute_data(
  nc_file,
  att_data,
  instance_dim_name = "instance",
  units = rep("unknown", ncol(att_data)),
  overwrite = FALSE
)

Arguments

nc_file
character file path to the nc file to be created. If adding to a file, it must already have the named instance dimension.

att_data
data.frame with instances as columns and attributes as rows.

instance_dim_name
character name for the instance dimension. Defaults to "instance"

units
character vector with units for each column of att_data. Defaults to "unknown" for all.

overwrite
boolean overwrite existing file? Will append if FALSE.

Examples

write_attribute_data,Write attribute data to NetCDF-CF,Write attribute data to NetCDF-CF — write_attribute_data,Source: R/write_attribute_data.R,write_attribute_data.Rd,character(0),write_attribute_data,character(0),list(title = "Description", contents = "
Creates a NetCDF file with an instance dimension, and any attributes from a data frame. \nUse to create the start of a NetCDF-DSG file. One character length dimension is created\nlong enough to contain the longest provided character string.\nThis function does not implement any CF convention attributes or standard names.\nAny columns of class date will be converted to character.
"),list(description = "Creates a NetCDF file with an instance dimension, and any attributes from a data frame. \nUse to create the start of a NetCDF-DSG file. One character length dimension is created\nlong enough to contain the longest provided character string.\nThis function does not implement any CF convention attributes or standard names.\nAny columns of class date will be converted to character."),write_attribute_data(
  nc_file,
  att_data,
  instance_dim_name = "instance",
  units = rep("unknown", ncol(att_data)),
  overwrite = FALSE
),list(list(name = "nc_file", description = "
character file path to the nc file to be created. \nIf adding to a file, it must already have the named instance dimension.
"), list(name = "att_data", description = "
data.frame with instances as columns and attributes as rows.
"), list(name = "instance_dim_name", description = "
character name for the instance dimension. Defaults to \"instance\"
"), list(name = "units", description = "
character vector with units for each column of att_data. Defaults to \"unknown\" for all.
"), 
    list(name = "overwrite", description = "
boolean overwrite existing file? Will append if FALSE.
")),TRUE,
sample_data <- sf::st_set_geometry(sf::read_sf(system.file("shape/nc.shp", 
                                                           package = "sf")), 
                                   NULL)
example_file <-write_attribute_data(tempfile(), sample_data,
                                    units = rep("unknown", ncol(sample_data)))

try({
  ncdump <- system(paste("ncdump -h", example_file), intern = TRUE)
  cat(ncdump ,sep = "\n")
}, silent = TRUE)

#> netcdf file1051875a058b1 {
#> dimensions:
#> 	instance = 100 ;
#> 	char = 12 ;
#> variables:
#> 	double AREA(instance) ;
#> 		AREA:units = "unknown" ;
#> 		AREA:missing_value = -9999.999 ;
#> 	double PERIMETER(instance) ;
#> 		PERIMETER:units = "unknown" ;
#> 		PERIMETER:missing_value = -9999.999 ;
#> 	double CNTY_(instance) ;
#> 		CNTY_:units = "unknown" ;
#> 		CNTY_:missing_value = -9999.999 ;
#> 	double CNTY_ID(instance) ;
#> 		CNTY_ID:units = "unknown" ;
#> 		CNTY_ID:missing_value = -9999.999 ;
#> 	char NAME(instance, char) ;
#> 		NAME:units = "unknown" ;
#> 	char FIPS(instance, char) ;
#> 		FIPS:units = "unknown" ;
#> 	double FIPSNO(instance) ;
#> 		FIPSNO:units = "unknown" ;
#> 		FIPSNO:missing_value = -9999.999 ;
#> 	int CRESS_ID(instance) ;
#> 		CRESS_ID:units = "unknown" ;
#> 		CRESS_ID:missing_value = -9999 ;
#> 	double BIR74(instance) ;
#> 		BIR74:units = "unknown" ;
#> 		BIR74:missing_value = -9999.999 ;
#> 	double SID74(instance) ;
#> 		SID74:units = "unknown" ;
#> 		SID74:missing_value = -9999.999 ;
#> 	double NWBIR74(instance) ;
#> 		NWBIR74:units = "unknown" ;
#> 		NWBIR74:missing_value = -9999.999 ;
#> 	double BIR79(instance) ;
#> 		BIR79:units = "unknown" ;
#> 		BIR79:missing_value = -9999.999 ;
#> 	double SID79(instance) ;
#> 		SID79:units = "unknown" ;
#> 		SID79:missing_value = -9999.999 ;
#> 	double NWBIR79(instance) ;
#> 		NWBIR79:units = "unknown" ;
#> 		NWBIR79:missing_value = -9999.999 ;
#> }

,list(),2021,list(name = "ncdfgeom", version = "1.1.1", authors = "David Blodgett"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "ncdfgeom"),list(),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Get started\n
\n
\n  Reference\n
\n
\n  \n    Articles\n     \n    \n  \n  
\n    
\n      Reading and Writing NetCDF-CF Geometry\n    
\n    
\n      Reading and Writing Timeseries\n    
\n  
\n
\n
\n  Changelog\n
", 
    right = "
\n  \n    \n     \n  \n
"),list(.present = TRUE),list(version = "1.6.1"),FALSE