Write attribute data to NetCDF-CF

Creates a NetCDF file with an instance dimension, and any attributes from a data frame. Use to create the start of a NetCDF-DSG file. One character length dimension is created long enough to contain the longest provided character string. This function does not implement any CF convention attributes or standard names. Any columns of class date will be converted to character.

write_attribute_data(nc_file, attData, instance_dim_name = "instance",
  units = rep("unknown", ncol(attData)), ...)

Arguments

nc_file
character file path to the nc file to be created. If adding to a file, it must already have the named instance dimension.

attData
data.frame with instances as rows and attributes as rows.

instance_dim_name
character name for the instance dimension. Defaults to "instance"

units
character vector with units for each column of attData. Defaults to "unknown" for all.

...
additional arguments to be passed on to nc_create.

Examples

sample_data <- sf::st_set_geometry(sf::read_sf(system.file("shape/nc.shp",
                                                           package = "sf")),
                                   NULL)
example_file <-write_attribute_data(tempfile(), sample_data,
                                    units = rep("unknown", ncol(sample_data)))

ncdump <- system(paste("ncdump -h", example_file), intern = TRUE)
cat(ncdump ,sep = "\n")

#> netcdf filea3d02befb82f {
#> dimensions:
#> 	instance = 100 ;
#> 	char = 12 ;
#> variables:
#> 	double AREA(instance) ;
#> 		AREA:units = "unknown" ;
#> 		AREA:_FillValue = NaN ;
#> 	double PERIMETER(instance) ;
#> 		PERIMETER:units = "unknown" ;
#> 		PERIMETER:_FillValue = NaN ;
#> 	double CNTY_(instance) ;
#> 		CNTY_:units = "unknown" ;
#> 		CNTY_:_FillValue = NaN ;
#> 	double CNTY_ID(instance) ;
#> 		CNTY_ID:units = "unknown" ;
#> 		CNTY_ID:_FillValue = NaN ;
#> 	char NAME(instance, char) ;
#> 		NAME:units = "unknown" ;
#> 	char FIPS(instance, char) ;
#> 		FIPS:units = "unknown" ;
#> 	double FIPSNO(instance) ;
#> 		FIPSNO:units = "unknown" ;
#> 		FIPSNO:_FillValue = NaN ;
#> 	int CRESS_ID(instance) ;
#> 		CRESS_ID:units = "unknown" ;
#> 		CRESS_ID:_FillValue = -9999 ;
#> 	double BIR74(instance) ;
#> 		BIR74:units = "unknown" ;
#> 		BIR74:_FillValue = NaN ;
#> 	double SID74(instance) ;
#> 		SID74:units = "unknown" ;
#> 		SID74:_FillValue = NaN ;
#> 	double NWBIR74(instance) ;
#> 		NWBIR74:units = "unknown" ;
#> 		NWBIR74:_FillValue = NaN ;
#> 	double BIR79(instance) ;
#> 		BIR79:units = "unknown" ;
#> 		BIR79:_FillValue = NaN ;
#> 	double SID79(instance) ;
#> 		SID79:units = "unknown" ;
#> 		SID79:_FillValue = NaN ;
#> 	double NWBIR79(instance) ;
#> 		NWBIR79:units = "unknown" ;
#> 		NWBIR79:_FillValue = NaN ;
#> }