Write NetCDF-CF point featuretype

Creates a point feature type discrete sampling features NetCDF file. Returns the created filename. Can pass in netcdf creation options like force_v4 to pass on to nc_create().

write_point_dsg(nc_file, lats, lons, alts, times, data,
  data_units = rep("", ncol(data)), feature_names = NULL, ...)

Arguments

nc_file
character file path to the nc file to be created.

lats
numeric vector of latitudes

lons
numeric vector of longitudes

alts
numeric vector of altitudes

times
POSIXct vector of times, one per point, if length is one, will use the same time for each point. Must be of type POSIXct or an attempt to convert it will be made using as.POSIXct(times).

data
data.frame with each column corresponding to a observation. Column names are used as names in the NCDF file

data_units
character vector of observation units. Length must be the same as number of columns in data parameter

feature_names
character or numeric vector of identifiers for features or stations.

...
additional arguments to be passed on nc_create.

Value

character path of created file

References

http://cfconventions.org/cf-conventions/cf-conventions.html#_features_and_feature_types

http://www.unidata.ucar.edu/software/thredds/current/netcdf-java/reference/FeatureDatasets/CFpointImplement.html