Create timeseries NCDF file

This creates a timeseries discrete sampling geometry NCDF file

write_timeseries_dsg(nc_file, station_names, lats, lons, times, data,
  alts = NA, data_unit = "", data_prec = "double",
  data_metadata = list(name = "data", long_name = "unnamed data"),
  attributes = list(), add_to_existing = FALSE)

Arguments

nc_file
A string file path to the nc file to be created.

station_names
A vector of names for each station to be added to the file.

lats
Vector of latitudes

lons
Vector of longitudes

times
A vector of times. Must be of type POSIXct or an attempt to convert it will be made using as.POSIXct(times).

data
data.frame with each column corresponding to a station. Rows correspond to time steps. nrow must be the same length as times. Column names must match station names.

alts
Vector of altitudes (Optional)

data_unit
Character vector of observation units. Length must be the same as number of columns in data parameter

data_prec
Precision of observation data in NCDF file. Valid options: 'short' 'integer' 'float' 'double' 'char'.

data_metadata
A named list of strings: list(name='ShortVarName', long_name='A Long Name')

attributes
An optional list of attributes that will be added at the global level. See details for useful attributes.

add_to_existing
boolean If TRUE and the file already exists, variables will be added to the existing file. Note that this should be used to add variables not stations or time steps. All inputs should be the same as are already in the file.

Details

title = "title" abstract = "history" provider site = "institution" provider name ="source" description = "description"

References

http://www.unidata.ucar.edu/software/thredds/current/netcdf-java/reference/FeatureDatasets/CFpointImplement.html